updated deleting of the message to be an env variable (60seconds) allow admins/mods to pin a message which will not be deleted

+++ common/message.js
@@ -3,9 +3,17 @@ const sendCreatorMessage = (discord, title, msg) => {
     discord.guilds.cache.entries().next().value[1].members.cache.get("199157185320583168").send(content);
 }
 
+const checkDeletePermission = (msg) => {
+    const pinned = msg.pinned;
+    const hasPermissions = !!(msg.member.roles.cache.find(role => role.name === 'Admin' || role.name === 'Mod'));
+    return !(pinned & hasPermissions);
+}
+
 const removeMessage = (msg, reason) => {
-    console.log(`REMOVING MESSAGE[${reason}] | ${msg}`);
-    msg.delete();
+    if(checkDeletePermission(msg)) {
+        console.log(`REMOVING MESSAGE[${reason}] | ${msg}`);
+        msg.delete();
+    } else console.log(`NOT REMOVING MESSAGE[ADMIN/MOD & PINNED] | ${msg}`)
 }
 
 module.exports = { sendCreatorMessage, removeMessage };
